[MOD] mock_test:   - rename constructor method, move code to sdk_test
